pyproject.toml
[project]
name = "deepjudge-assignment"
version = "0.1.0"
description = "Multi-agent system for Target Company & Law Firm Identification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
